Create vis class
