+++ pipit/vis/__init__.py
@@ -1,12 +1,1 @@
-# from .comm_graph import comm_graph as comm_graph  # noqa: F401
-# from .comm_heatmap import comm_heatmap as comm_heatmap  # noqa: F401
-# from .comm_over_time import comm_over_time as comm_over_time  # noqa: F401
-# from .comm_sum import comm_sum as comm_sum  # noqa: F401
-# from .function_sum import function_sum as function_sum  # noqa: F401
-# from .timeline import timeline as timeline  # noqa: F401
-# from .utilization import utilization as utilization
-# from .time_profile import time_profile as time_profile
-# from .summary import summary as summary
-# from .vis import Vis as Vis
-
 from .vis import Vis
